# docs/swagger.yaml
basePath: /
definitions:
...
  controller.Info:
    properties:
      name:
        type: string
      price:
        type: integer
    type: object
  ds.Consultation:
    properties:
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      price:
        type: integer
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Consultation app
  title: IT Services
  version: "1.0"
paths:
...
  /consultation-request/delete/consultation/{id_c}/request/{id_r}:
    delete:
      consumes:
      - application/json
      description: delete consultation from request
      operationId: delete-consultation-from-request
      parameters:
      - description: ID консультации
        in: path
        name: id_c
        required: true
        type: integer
      - description: ID заявки
        in: path
        name: id_r
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Консультация была удалена из заявки
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            type: string
        "404":
          description: Некорректный запрос
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete Consultation From Request
      tags:
      - Consultation-Request
  /consultations:
    get:
      description: Get all consultations
      operationId: get-consultations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.Consultation'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Consultation'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Consultation'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Consultation'
      summary: Get Consultations
      tags:
      - Consultations
  /consultations/{id}:
    get:
      consumes:
      - application/json
      description: Show consultation by ID
      operationId: get-consultation-by-id
      parameters:
      - description: ID консультации
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.Consultation'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Consultation'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Consultation'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Consultation'
      summary: Get Consultation by ID
      tags:
      - Consultations
  /consultations/{id}/add-to-request:
    post:
      consumes:
      - application/json
      description: add consultation to request
      operationId: add-consultation-to-request
      parameters:
      - description: ID консультации
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Consultation'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Consultation'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Consultation'
      security:
      - ApiKeyAuth: []
      summary: add consultation to request
      tags:
      - Consultations
  /consultations/{id}/addImage:
    post:
      consumes:
      - application/json
      description: Add an image to a specific consultation by ID.
      operationId: add-consultation-image
      parameters:
      - description: ID консультации
        in: path
        name: id
        required: true
        type: integer
      - description: Image file to be uploaded
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Consultation'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Consultation'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Consultation'
      security:
      - ApiKeyAuth: []
      summary: Add consultation image
      tags:
      - Consultations
  /consultations/create:
    post:
      consumes:
      - application/json
      description: create consultation
      operationId: create-consultation
      parameters:
      - description: consultation info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.Consultation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Consultation'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Consultation'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Consultation'
      security:
      - ApiKeyAuth: []
      summary: create consultation
      tags:
      - Consultations
  /consultations/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete consultation by ID
      operationId: delete-consultation-by-id
      parameters:
      - description: ID консультации
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Consultation'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Consultation'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Consultation'
      summary: Delete consultation by ID
      tags:
      - Consultations
  /consultations/request/{id}:
    get:
      consumes:
      - application/json
      description: Show consultation by ID of request
      operationId: get-consultation-by-id-of-request
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Info'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Consultation'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Consultation'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Consultation'
      security:
      - ApiKeyAuth: []
      summary: Get Consultation by request ID
      tags:
      - Consultations
  /consultations/update/{id}:
    put:
      consumes:
      - application/json
      description: update consultation
      operationId: update-consultation
      parameters:
      - description: ID консультации
        in: path
        name: id
        required: true
        type: integer
      - description: consultation info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.Consultation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Consultation'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Consultation'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Consultation'
      security:
      - ApiKeyAuth: []
      summary: update consultation
      tags:
      - Consultations
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"